refactor(query): extract phase fs and roadmap mutation adapters
